requirements.txt
fastapi
uvicorn
gitpython
rich
